packages/forwardref_inspect/forwardref_inspect-0.1.0.tar.gz/forwardref_inspect-0.1.0/tests/test_inspect.py
import forwardref_inspect

from tests.foo import Foo


def test_forward_reference_str_parameters_annotation():
    signature = forwardref_inspect.signature(Foo.__eq__)
    assert signature.parameters['other'].annotation == Foo


def test_forward_reference_mixed_parameters_annotation():
    signature = forwardref_inspect.signature(Foo.method_with_mixed_annotations)
    assert signature.parameters['other'].annotation == Foo
    assert signature.parameters['flag'].annotation == bool


def test_forward_reference_str_return_annotation():
    signature = forwardref_inspect.signature(Foo.build)
    assert signature.return_annotation == Foo
